feat: final result

+++ src/csv_util.h
@@ -14,7 +14,7 @@
 
 // append a line of data to a csv format file
 // the object label is written to the first column
-// the object features are all written to the following columns as floats
+// the object features are all written to the following columns as doubles
 // 
 // filename: the name of the file to append to
 // object_label: the label for the object
@@ -22,7 +22,7 @@
 // reset_file: if true, then the file is opened in 'write' mode and the existing contents are cleared.
 // if false, then the file is opened in 'append' mode and the new data is appended to the end of the file.
 // returns a non-zero value if something goes wrong
-int append_object_data_csv(std::string filename, std::string object_label, std::vector<float> &object_feature, int reset_file = 0);
+int append_object_data_csv(std::string filename, std::string object_label, std::vector<double> &object_feature, int reset_file = 0);
 
 // read data from a csv format file
 // whose first column is the object label and the remaining columns are the object features
@@ -30,9 +30,9 @@ int append_object_data_csv(std::string filename, std::string object_label, std::
 // 
 // filename: the name of the file to read from
 // object_labels: the labels for the objects as a std::vector of std::string
-// object_features: the features for the objects as a 2D std::vector<float>
+// object_features: the features for the objects as a 2D std::vector<double>
 // echo_file: if true, then the contents of the file are printed to the console
 // returns a non-zero value if something goes wrong
-int read_object_data_csv(std::string filename, std::vector<std::string> &object_labels, std::vector<std::vector<float>> &object_features, int echo_file = 0);
+int read_object_data_csv(std::string filename, std::vector<std::string> &object_labels, std::vector<std::vector<double>> &object_features, int echo_file = 0);
 
 #endif
